AddTile
created addTile component- basically what will be the base for actually adding csv's to the app to be able to see them.
played around with the modal part, it pops up now. and is a thing in general ig. it'll be the actual thing that pops up when the add tile is pressed to be able to find a new csv to upload and edit the tile's display name and potentially color